+++ packages/uploader-server/src/utils.ts
@@ -15,8 +15,13 @@ export let logger = console;
 
 // it means we are running in blocklet environment, use logger from @blocklet/logger
 if (process.env.BLOCKLET_LOG_DIR) {
-  const initLogger = require('@blocklet/logger');
-  logger = initLogger('uploader-server');
+  try {
+    const initLogger = require('@blocklet/logger');
+    logger = initLogger('uploader-server');
+    logger.info('uploader-server logger init success');
+  } catch (error) {
+    // ignore
+  }
 }
 
 // fork from @blocklet/sdk/lib/component/index.d.ts